+++ mpapt-runtime/src/main/java/de/jensklingenberg/mpapt/model/AbstractProcessor.kt
@@ -27,7 +27,7 @@ abstract class AbstractProcessor() : Processor {
     fun log(message: String) {
         messageCollector().report(
                 CompilerMessageSeverity.WARNING,
-                "*** AbstractProcessor: $message"
+                 message
         )
     }